backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.31.0
python-dotenv>=0.19.0
Werkzeug>=2.0.0 
python-amazon-paapi>=5.0.0
beautifulsoup4>=4.9.3
pandas>=1.3.0
numpy>=1.20.0
pillow>=8.0.0
google-cloud-vision>=3.1.0
google-api-python-client>=2.0.0
lxml>=4.6.3